+++ persistence/module.go
@@ -3,8 +3,9 @@ package persistence
 import (
 	"context"
 	"fmt"
-	"github.com/pokt-network/pocket/persistence/indexer"
 	"log"
+
+	"github.com/pokt-network/pocket/persistence/indexer"
 
 	"github.com/pokt-network/pocket/persistence/types"
 
@@ -37,7 +38,7 @@ type persistenceModule struct {
 }
 
 const (
-	PersistenceModuleName = "persistence"
+	persistenceModuleName = "persistence"
 )
 
 func Create(runtimeMgr modules.RuntimeMgr) (modules.Module, error) {
@@ -115,7 +116,7 @@ func (m *persistenceModule) Stop() error {
 }
 
 func (m *persistenceModule) GetModuleName() string {
-	return PersistenceModuleName
+	return persistenceModuleName
 }
 
 func (m *persistenceModule) SetBus(bus modules.Bus) {
@@ -130,10 +131,12 @@ func (m *persistenceModule) GetBus() modules.Bus {
 }
 
 func (*persistenceModule) ValidateConfig(cfg modules.Config) error {
+	// TODO (#334): implement this
 	return nil
 }
 
 func (*persistenceModule) ValidateGenesis(genesis modules.GenesisState) error {
+	// TODO (#334): implement this
 	return nil
 }
 
@@ -189,6 +192,10 @@ func (m *persistenceModule) NewReadContext(height int64) (modules.PersistenceRea
 	}, nil
 }
 
+func (m *persistenceModule) ReleaseWriteContext() error {
+	panic("TODO(#284): Implement proper write context release.")
+}
+
 func (m *persistenceModule) GetBlockStore() kvstore.KVStore {
 	return m.blockStore
 }
@@ -204,8 +211,8 @@ func initializeBlockStore(blockStorePath string) (kvstore.KVStore, error) {
 	return kvstore.NewKVStore(blockStorePath)
 }
 
-// TODO(drewsky): Simplify and externalize the logic for whether genesis should be populated and
-// move the if logic out of this file.
+// HACK(olshansky): Simplify and externalize the logic for whether genesis should be populated and
+//                  move the if logic out of this file.
 func (m *persistenceModule) shouldHydrateGenesisDb() (bool, error) {
 	checkContext, err := m.NewReadContext(-1)
 	if err != nil {
@@ -213,10 +220,8 @@ func (m *persistenceModule) shouldHydrateGenesisDb() (bool, error) {
 	}
 	defer checkContext.Close()
 
-	maxHeight, err := checkContext.GetLatestBlockHeight()
-	if err == nil || maxHeight == 0 {
+	if _, err = checkContext.GetLatestBlockHeight(); err != nil {
 		return true, nil
 	}
-
-	return m.blockStore.Exists(heightToBytes(int64(maxHeight)))
+	return false, nil
 }