+++ shared/bus.go
@@ -3,7 +3,7 @@ package shared
 import (
 	"log"
 
-	"github.com/pokt-network/pocket/shared/debug"
+	"github.com/pokt-network/pocket/shared/messaging"
 	"github.com/pokt-network/pocket/shared/modules"
 )
 
@@ -21,6 +21,8 @@ type bus struct {
 	utility     modules.UtilityModule
 	consensus   modules.ConsensusModule
 	telemetry   modules.TelemetryModule
+	logger      modules.LoggerModule
+	rpc         modules.RPCModule
 
 	runtimeMgr modules.RuntimeMgr
 }
@@ -36,6 +38,8 @@ func CreateBus(
 	utility modules.UtilityModule,
 	consensus modules.ConsensusModule,
 	telemetry modules.TelemetryModule,
+	logger modules.LoggerModule,
+	rpc modules.RPCModule,
 ) (modules.Bus, error) {
 	bus := &bus{
 		channel: make(modules.EventsChannel, DefaultPocketBusBufferSize),
@@ -47,6 +51,8 @@ func CreateBus(
 		utility:     utility,
 		consensus:   consensus,
 		telemetry:   telemetry,
+		logger:      logger,
+		rpc:         rpc,
 	}
 
 	modules := map[string]modules.Module{
@@ -55,6 +61,8 @@ func CreateBus(
 		"p2p":         p2p,
 		"utility":     utility,
 		"telemetry":   telemetry,
+		"logger":      logger,
+		"rpc":         rpc,
 	}
 
 	// checks if modules are not nil and sets their bus to this bus instance.
@@ -86,6 +94,8 @@ func CreateBusWithOptionalModules(
 	utility modules.UtilityModule,
 	consensus modules.ConsensusModule,
 	telemetry modules.TelemetryModule,
+	logger modules.LoggerModule,
+	rpc modules.RPCModule,
 ) modules.Bus {
 	bus := &bus{
 		channel: make(modules.EventsChannel, DefaultPocketBusBufferSize),
@@ -97,6 +107,8 @@ func CreateBusWithOptionalModules(
 		utility:     utility,
 		consensus:   consensus,
 		telemetry:   telemetry,
+		logger:      logger,
+		rpc:         rpc,
 	}
 
 	maybeSetModuleBus := func(mod modules.Module) {
@@ -110,15 +122,17 @@ func CreateBusWithOptionalModules(
 	maybeSetModuleBus(utility)
 	maybeSetModuleBus(consensus)
 	maybeSetModuleBus(telemetry)
+	maybeSetModuleBus(logger)
+	maybeSetModuleBus(rpc)
 
 	return bus
 }
 
-func (m bus) PublishEventToBus(e *debug.PocketEvent) {
+func (m bus) PublishEventToBus(e *messaging.PocketEnvelope) {
 	m.channel <- *e
 }
 
-func (m bus) GetBusEvent() *debug.PocketEvent {
+func (m bus) GetBusEvent() *messaging.PocketEnvelope {
 	e := <-m.channel
 	return &e
 }
@@ -145,6 +159,14 @@ func (m bus) GetConsensusModule() modules.ConsensusModule {
 
 func (m bus) GetTelemetryModule() modules.TelemetryModule {
 	return m.telemetry
+}
+
+func (m bus) GetLoggerModule() modules.LoggerModule {
+	return m.logger
+}
+
+func (m bus) GetRPCModule() modules.RPCModule {
+	return m.rpc
 }
 
 func (m *bus) GetRuntimeMgr() modules.RuntimeMgr {