+++ shared/messaging/events.go
@@ -1,0 +1,5 @@
+package messaging
+
+const (
+	NodeStartedEventType = "pocket.NodeStartedEvent"
+)