+++ consensus/doc/CHANGELOG.md
@@ -7,8 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.0.8] - 2022-11-09
+## [0.0.0.9] - 2022-11-30
 - Added state sync interfaces and diagrams
+
+## [0.0.0.8] - 2022-11-15
+
+- Propagate the `quorumCertificate` on `Block` commit to the `Utility` module
+- Slightly improved error handling of the `utilityContext` lifecycle management
 
 ## [0.0.0.7] - 2022-11-01