+++ p2p/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0.8] - 2022-11-14
+
+- Removed topic from messaging
+
 ## [0.0.0.7] - 2022-10-24
 
 - Updated README to reference the python simulator as a learning references and unit test generation tool